backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
jinja2==3.1.3
weasyprint==60.2
python-docx==1.1.0
pillow==10.2.0
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
python-dotenv==1.0.1
